+++ src/graphics/Drawables.h
@@ -143,8 +143,8 @@ private:
 	//add a new vertex, return the index
 	int AddVertex(VertexArray&, const vector3f &v, const vector3f &n);
 	//add three vertex indices to form a triangle
-	void AddTriangle(std::vector<Uint16>&, int i1, int i2, int i3);
-	void Subdivide(VertexArray&, std::vector<Uint16>&,
+	void AddTriangle(std::vector<Uint32>&, int i1, int i2, int i3);
+	void Subdivide(VertexArray&, std::vector<Uint32>&,
 		const matrix4x4f &trans, const vector3f &v1, const vector3f &v2, const vector3f &v3,
 		int i1, int i2, int i3, int depth);
 };
